Added column checks to Alembic

+++ alembic/versions/de5d615ae090_add_user_to_penalty_stat.py
@@ -5,9 +5,13 @@ Revises: 31918b83c372
 Create Date: 2022-10-14 19:33:02.808038
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
+from sqlalchemy.engine.reflection import Inspector
 
+conn = op.get_bind()
+inspector = Inspector.from_engine(conn)
+tables = inspector.get_table_names()
 
 # revision identifiers, used by Alembic.
 revision = "de5d615ae090"
@@ -16,17 +20,33 @@ branch_labels = None
 depends_on = None
 
 
+def _table_has_column(table, column):
+    has_column = False
+    for col in inspector.get_columns(table):
+        if column not in col["name"]:
+            continue
+        has_column = True
+    return has_column
+
+
+def _has_table(table_name):
+    tables = inspector.get_table_names()
+    return table_name in tables
+
+
 def upgrade():
-    op.add_column("penalty", sa.Column("user_id", sa.INTEGER))
-    op.create_foreign_key(
-        "penalty_ibfk_3",
-        "penalty",
-        "user",
-        ["user_id"],
-        ["id"],
-        ondelete="SET NULL",
-    )
+    if not _table_has_column("penalty", "user_id"):
+        op.add_column("penalty", sa.Column("user_id", sa.INTEGER))
+        op.create_foreign_key(
+            "penalty_ibfk_3",
+            "penalty",
+            "user",
+            ["user_id"],
+            ["id"],
+            ondelete="SET NULL",
+        )
 
 
 def downgrade():
-    op.drop_column("penalty", "user_id")
+    if _table_has_column("penalty", "user_id"):
+        op.drop_column("penalty", "user_id")
